Enable additional low-noise golangci-lint linters
Enable a set of correctness guards and small cleanup/modernization
linters that are either already clean on the tree or require only
mechanical fixes:

  errname, usestdlibvars, nilerr, bodyclose, durationcheck, makezero,
  reassign, predeclared, wastedassign, unconvert, whitespace, intrange,
  copyloopvar, misspell

errname reinforces the existing sentinel-error naming convention; the
correctness linters (nilerr, bodyclose, durationcheck, makezero,
reassign, predeclared, wastedassign) are clean today and guard against
future regressions.

Fixes applied for the findings this surfaced:

  - errname: rename noIdentityError -> errNoIdentity, fakeTimeoutErr ->
    fakeTimeoutError, and the osStatus error type -> osStatusError (the
    former osStatusError constructor becomes newOSStatusError).
  - unconvert: drop redundant conversions of already-typed constants
    (jceks entry tags, tls cipher suites, proxyproto command/protocol).
  - usestdlibvars: use http.Method* constants; add the net/http import
    where the rewrite required it.
  - intrange/copyloopvar/whitespace/misspell: mechanical modernizations
    and cleanups applied via golangci-lint --fix.

+++ certloader/jceks/jceks_test.go
@@ -61,7 +61,7 @@ func buildMinimalJCEKS(t *testing.T, alias string, certDER []byte, password stri
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(1)))
 
 	// Trusted cert entry tag
-	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(trustedCertEntryTag)))
+	require.NoError(t, binary.Write(&body, binary.BigEndian, trustedCertEntryTag))
 
 	// Alias (modified UTF-8 string with 2-byte length prefix)
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint16(len(alias))))
@@ -217,7 +217,7 @@ func buildJCEKSWithPrivateKey(t *testing.T, alias string, encryptedKeyDER []byte
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(1)))
 
 	// Private key entry tag
-	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(privateKeyEntryTag)))
+	require.NoError(t, binary.Write(&body, binary.BigEndian, privateKeyEntryTag))
 
 	// Alias
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint16(len(alias))))
@@ -804,7 +804,7 @@ func buildJCEKSWithPrivateKeyNoCerts(t *testing.T, alias string, encryptedKeyDER
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(jceksMagic)))
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(jceksVersion)))
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(1)))
-	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(privateKeyEntryTag)))
+	require.NoError(t, binary.Write(&body, binary.BigEndian, privateKeyEntryTag))
 
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint16(len(alias))))
 	_, err = body.WriteString(alias)
@@ -950,7 +950,7 @@ func TestParsePrivateKeyTruncatedAlias(t *testing.T) {
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(jceksMagic)))
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(jceksVersion)))
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(1)))
-	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(privateKeyEntryTag)))
+	require.NoError(t, binary.Write(&body, binary.BigEndian, privateKeyEntryTag))
 	// EOF here — no alias
 
 	h := MakeIntegrityHash(encodedPassword)
@@ -971,7 +971,7 @@ func TestParsePrivateKeyTruncatedDate(t *testing.T) {
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(jceksMagic)))
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(jceksVersion)))
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(1)))
-	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(privateKeyEntryTag)))
+	require.NoError(t, binary.Write(&body, binary.BigEndian, privateKeyEntryTag))
 	// Alias
 	alias := "a"
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint16(len(alias))))
@@ -996,7 +996,7 @@ func TestParseTrustedCertTruncatedDate(t *testing.T) {
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(jceksMagic)))
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(jceksVersion)))
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(1)))
-	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(trustedCertEntryTag)))
+	require.NoError(t, binary.Write(&body, binary.BigEndian, trustedCertEntryTag))
 	alias := "a"
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint16(len(alias))))
 	body.WriteString(alias)
@@ -1020,7 +1020,7 @@ func TestParseTrustedCertTruncatedCert(t *testing.T) {
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(jceksMagic)))
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(jceksVersion)))
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(1)))
-	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(trustedCertEntryTag)))
+	require.NoError(t, binary.Write(&body, binary.BigEndian, trustedCertEntryTag))
 	alias := "a"
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint16(len(alias))))
 	body.WriteString(alias)
@@ -1248,7 +1248,7 @@ func TestParseJKSMagic(t *testing.T) {
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(jksMagic)))
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(jceksVersion)))
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(1)))
-	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(trustedCertEntryTag)))
+	require.NoError(t, binary.Write(&body, binary.BigEndian, trustedCertEntryTag))
 	alias := "myalias"
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint16(len(alias))))
 	body.WriteString(alias)
@@ -1306,7 +1306,7 @@ func TestParsePrivateKeyTruncatedProtectedKey(t *testing.T) {
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(jceksMagic)))
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(jceksVersion)))
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(1)))
-	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(privateKeyEntryTag)))
+	require.NoError(t, binary.Write(&body, binary.BigEndian, privateKeyEntryTag))
 	alias := "a"
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint16(len(alias))))
 	body.WriteString(alias)
@@ -1331,7 +1331,7 @@ func TestParsePrivateKeyTruncatedCertCount(t *testing.T) {
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(jceksMagic)))
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(jceksVersion)))
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(1)))
-	require.NoError(t, binary.Write(&body, binary.BigEndian, uint32(privateKeyEntryTag)))
+	require.NoError(t, binary.Write(&body, binary.BigEndian, privateKeyEntryTag))
 	alias := "a"
 	require.NoError(t, binary.Write(&body, binary.BigEndian, uint16(len(alias))))
 	body.WriteString(alias)
@@ -1462,7 +1462,7 @@ func TestParseTrustedCertTruncatedAlias(t *testing.T) {
 	require.NoError(t, binary.Write(&buf, binary.BigEndian, uint32(jceksMagic)))
 	require.NoError(t, binary.Write(&buf, binary.BigEndian, uint32(jceksVersion)))
 	require.NoError(t, binary.Write(&buf, binary.BigEndian, uint32(1)))
-	require.NoError(t, binary.Write(&buf, binary.BigEndian, uint32(trustedCertEntryTag)))
+	require.NoError(t, binary.Write(&buf, binary.BigEndian, trustedCertEntryTag))
 	// EOF — readString fails reading the 2-byte length prefix.
 
 	ks := new(KeyStore)
@@ -1477,7 +1477,7 @@ func TestParsePrivateKeyTruncatedPerCert(t *testing.T) {
 	require.NoError(t, binary.Write(&buf, binary.BigEndian, uint32(jceksMagic)))
 	require.NoError(t, binary.Write(&buf, binary.BigEndian, uint32(jceksVersion)))
 	require.NoError(t, binary.Write(&buf, binary.BigEndian, uint32(1)))
-	require.NoError(t, binary.Write(&buf, binary.BigEndian, uint32(privateKeyEntryTag)))
+	require.NoError(t, binary.Write(&buf, binary.BigEndian, privateKeyEntryTag))
 	alias := "k"
 	require.NoError(t, binary.Write(&buf, binary.BigEndian, uint16(len(alias))))
 	buf.WriteString(alias)
